moved AI handler into enemy saga file
